requirements.txt

fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
httpx>=0.26.0

# AI & LLM
langchain>=0.1.0
langchain-google-genai>=0.0.6
langgraph>=0.0.20
crewai>=0.1.0

# Monitoring
langsmith>=0.0.87
